+++ src/libprojectM/MilkdropPresetFactory/Parser.cpp
@@ -2692,152 +2692,3 @@ else
 	return false;
 
 }
-
-
-
-
-
-
-
-// TESTS
-
-
-#include <TestRunner.hpp>
-
-#ifndef NDEBUG
-
-#include <PresetLoader.hpp>
-
-#define TEST(cond) if (!verify(#cond,cond)) return false
-#define TEST2(str,cond) if (!verify(str,cond)) return false
-
-struct ParserTest : public Test
-{
-    ParserTest() : Test("ParserTest")
-    {}
-
-    MilkdropPreset *preset;
-    std::istringstream is;
-    std::istringstream &ss(const char *s) { return is = std::istringstream(s); }
-
-    bool eq(float a, float b)
-    {
-        return std::abs(a-b) < (std::abs(a)+std::abs(b) + 1)/1000.0f;
-    }
-
-public:
-
-    bool test_float()
-    {
-        float f=-1.0f;
-        TEST(PROJECTM_SUCCESS == Parser::parse_float(ss("1.1"),&f));
-        TEST(1.1f == f);
-        TEST(PROJECTM_SUCCESS == Parser::parse_float(ss("+1.2"),&f));
-        TEST(PROJECTM_SUCCESS == Parser::parse_float(ss("-1.3"),&f));
-        TEST(PROJECTM_PARSE_ERROR == Parser::parse_float(ss(""),&f));
-        TEST(PROJECTM_PARSE_ERROR == Parser::parse_float(ss("\n"),&f));
-        TEST(PROJECTM_PARSE_ERROR == Parser::parse_float(ss("+"),&f));
-        return true;
-    }
-
-    bool test_int()
-    {
-        int i=-1;
-        TEST(PROJECTM_SUCCESS == Parser::parse_int(ss("1"),&i));
-        TEST(1 == i);
-        TEST(PROJECTM_SUCCESS == Parser::parse_int(ss("+2"),&i));
-        TEST(PROJECTM_SUCCESS == Parser::parse_int(ss("-3"),&i));
-        TEST(PROJECTM_PARSE_ERROR == Parser::parse_int(ss(""),&i));
-        TEST(PROJECTM_PARSE_ERROR == Parser::parse_int(ss("\n"),&i));
-        TEST(PROJECTM_PARSE_ERROR == Parser::parse_int(ss("+"),&i));
-        return true;
-    }
-
-    bool eval_expr(float expected, const char *s)
-    {
-        float result;
-        Expr *expr_parse = Parser::parse_gen_expr(ss(s),nullptr,preset);
-        TEST(expr_parse != nullptr);
-        // Expr doesn't really expect to run 'non-optimized' expressions any longer
-        Expr *expr = Expr::optimize(expr_parse);
-        expr_parse = nullptr;
-        if (!ParserTest::eq(expected, result=expr->eval(-1,-1)))
-        {
-            std::cout << "failed: expected " << expected << " found " << result << std::endl;
-            return false;
-        }
-        Expr::delete_expr(expr);
-        return true;
-    }
-
-    bool test_eqn()
-    {
-        TEST(eval_expr(1.0f, "1.0"));
-        TEST(eval_expr(-1.0f, "-(1.0)"));         // unary`
-        TEST(eval_expr(0.5f, "5/10.000"));        // binary
-        TEST(eval_expr(1.0f, "sin(3.14159/2)"));
-				// emtpy expression evals to 0, NULL usually means error in this parser
-				TEST(eval_expr(0.0f, ""));
-				TEST(eval_expr(0.0f, " "));
-				TEST(eval_expr(0.0f, "\n"));
-				TEST(eval_expr(0.0f, " \n"));
-				preset->presetOutputs().rot = 0.99f;
-        TEST(eval_expr(0.99f, "rot"));
-
-				// random other stuff to parse
-				Parser::parse_gen_expr(ss("0.5 + 0.5*sin(q8*0.613 + 1);"),nullptr,preset);
-        return true;
-    }
-
-    // test multi-line expression, and multi-expression line
-    bool test_lines()
-    {
-        // TODO
-        return true;
-    }
-
-    bool test_params()
-    {
-        // TODO
-        return true;
-    }
-
-
-    bool _test()
-    {
-        bool success = true;
-        success &= test_float();
-        success &= test_int();
-        success &= test_eqn();
-        success &= test_lines();
-        success &= test_params();
-        return success;
-    }
-
-    bool test() override
-    {
-        // load IdlePreset
-        PresetLoader *presetLoader = new PresetLoader ( 400, 400, "" );
-        std::unique_ptr<Preset> preset_ptr = presetLoader->loadPreset("idle://Geiss & Sperl - Feedback (projectM idle HDR mix).milk");
-        preset = (MilkdropPreset *)preset_ptr.get();
-
-        bool success = _test();
-
-        delete presetLoader;
-        return success;
-    }
-};
-
-Test* Parser::test()
-{
-  return new ParserTest();
-}
-
-#else
-
-Test* Parser::test()
-{
-    return nullptr;
-}
-
-#endif
