--- src/libprojectM/PresetChooser.cpp
@@ -1,14 +1,0 @@
-//
-// C++ Implementation: PresetChooser
-//
-// Description:
-//
-//
-// Author: Carmelo Piccione <carmelo.piccione@gmail.com>, (C) 2007
-//
-// Copyright: See COPYING file that comes with this distribution
-//
-//
-
-#include "PresetChooser.hpp"
-
